proc: drop FuzzCallInjectionProtocol
Keep deterministic tests for premature RestoreRegisters. The random
protocol-register fuzzer did not find unknown bugs.
