+++ _scripts/fuzz_eval_expression.sh
@@ -1,8 +1,7 @@
 #!/usr/bin/env bash
 # Run FuzzEvalExpression setup, then either the seed corpus (seed) or a
-# timed continuous fuzz (fuzz). Fuzz mode also runs 5s continuous smokes for
-# FuzzEvalStackOps and FuzzCallInjectionProtocol. Intended for CI;
-# linux/amd64 only.
+# timed continuous fuzz (fuzz). Fuzz mode also runs a 5s smoke for
+# FuzzEvalStackOps. Intended for CI; linux/amd64 only.
 set -euo pipefail
 
 mode="${1:-}"
@@ -34,10 +33,6 @@ fuzz)
 
 	echo "fuzz_eval_expression: FuzzEvalStackOps (5s)"
 	go test ./pkg/proc -count=1 -run=NONE -fuzz=FuzzEvalStackOps \
-		-fuzztime=5s -fuzzminimizetime=0
-
-	echo "fuzz_eval_expression: FuzzCallInjectionProtocol (5s)"
-	go test ./pkg/proc -count=1 -run=NONE -fuzz=FuzzCallInjectionProtocol \
 		-fuzztime=5s -fuzzminimizetime=0
 	;;
 esac
